+++ src/classes/index.ts
@@ -2,10 +2,12 @@ import type CalendarManager from "./calendar/calendar-manager";
 import type MainApp from "./applications/main-app";
 import type SCController from "./s-c-controller";
 import ConfigurationApp from "./applications/configuration-app";
+import MigrationApp from "./applications/migration-app";
 
 export let CalManager: CalendarManager;
 export let MainApplication: MainApp;
 export let ConfigurationApplication: ConfigurationApp;
+export let MigrationApplication: MigrationApp;
 export let SC: SCController;
 
 /**
@@ -30,6 +32,14 @@ export function updateMainApplication(app: MainApp){
  */
 export function updateConfigurationApplication(app: ConfigurationApp){
     ConfigurationApplication = app;
+}
+
+/**
+ * Updates the global migration application to the passed in value, should only be called once
+ * @param app
+ */
+export function updateMigrationApplication(app: MigrationApp){
+    MigrationApplication = app;
 }
 
 /**