+++ src/classes/migrations/v1-to-v2.ts
@@ -1,9 +1,17 @@
 import {SettingNames} from "../../constants";
 import {GameSettings} from "../foundry-interfacing/game-settings";
 import Calendar from "../calendar";
+import UserPermissions from "../configuration/user-permissions";
+import {SC} from "../index";
 
-
+/**
+ * Class for handling the migration from Simple Calendar V1 to V2
+ */
 export default class V1ToV2{
+
+    /**
+     * Migrates the old calendar settings to the new calendar settings.
+     */
     public static runCalendarMigration() : Calendar | null {
         //Create a calendar configuration with the legacy settings
         const legacySettings: SimpleCalendar.CalendarData = {
@@ -28,5 +36,55 @@ export default class V1ToV2{
             isValid = true;
         }
         return isValid ? newCalendar : null;
+    }
+
+    /**
+     * Migrates the old permissions and general settings to the new global configuration
+     */
+    public static runGlobalConfigurationMigration(): boolean {
+        const oldGeneralConfig = <any>GameSettings.GetObjectSettings(SettingNames.GeneralConfiguration);
+        const oldTimeConfig = <any>GameSettings.GetObjectSettings(SettingNames.TimeConfiguration);
+
+        let perms = false, settings = false;
+
+        //Parse the old permissions and save them in the new format
+        if(oldGeneralConfig.hasOwnProperty('permissions') && oldGeneralConfig['permissions']){
+            const permissions = new UserPermissions();
+            permissions.loadFromSettings(oldGeneralConfig['permissions']);
+            permissions.changeActiveCalendar = {player: false, trustedPlayer: false, assistantGameMaster: false, users: undefined};
+
+            SC.globalConfiguration.permissions = permissions;
+            perms = true;
+        }
+
+        //Parse out the seconds per combat round and save them in the new format
+        if(oldTimeConfig.hasOwnProperty('secondsInCombatRound') && oldTimeConfig['secondsInCombatRound']){
+            const sICR = parseInt(oldTimeConfig['secondsInCombatRound'].toString());
+            if(!isNaN(sICR)){
+                SC.globalConfiguration.secondsInCombatRound = sICR;
+                settings = true;
+            }
+        }
+        return perms && settings;
+    }
+
+    public static runNoteMigration(): boolean {
+        return true;
+    }
+
+    public static async cleanUpOldData(): Promise<boolean> {
+        const gcRes = await GameSettings.SaveObjectSetting(SettingNames.GeneralConfiguration, {});
+        const ycRes = await GameSettings.SaveObjectSetting(SettingNames.YearConfiguration, {});
+        const wcRes = await GameSettings.SaveObjectSetting(SettingNames.WeekdayConfiguration, []);
+        const mcRes = await GameSettings.SaveObjectSetting(SettingNames.MonthConfiguration, []);
+        const cdRes = await GameSettings.SaveObjectSetting(SettingNames.CurrentDate, {});
+        const lyrRes = await GameSettings.SaveObjectSetting(SettingNames.LeapYearRule, {});
+        const tcRes = await GameSettings.SaveObjectSetting(SettingNames.TimeConfiguration, {});
+        const scRes = await GameSettings.SaveObjectSetting(SettingNames.SeasonConfiguration, []);
+        const moonRes = await GameSettings.SaveObjectSetting(SettingNames.MoonConfiguration, []);
+        const nRes = await GameSettings.SaveObjectSetting(SettingNames.Notes, []);
+        const ncRes = await GameSettings.SaveObjectSetting(SettingNames.NoteCategories, []);
+
+        return gcRes && ycRes && wcRes && mcRes && cdRes && lyrRes && tcRes && scRes && moonRes && nRes && ncRes;
     }
 }